Adds vue_provide templatetag for providing data to VueJS
Adds the `vue_provide` templatetag to enable the developer to embed
data in the rendered template for use by VueJS code. This makes use of
VueJS's `provide` and `inject` functionality for passing state/data
down the component stack without directly passing the data as _props_.

At this time, this functionality is made possible in the VueJS
codebase through Mike Hoolehan's `vue-plugin-django-utils` package.