Use v-model for form Select component
Use the v-model directive to simplify the custom Select
component. Provides a more intuitive feel as if one were simply using
a standard select element.

A watch of the condition's identifier value was put into the
ConditionRow component to enable the behavior that resets the value or
leaves it the same depending on the availablity of the identifier's
schema options. IMHO, this is now a more straight forward operation,
where the v-model logic does its behavior and we simply watch for the
change in order to make additional changes as needed.