Adds extraction of sticky conditions from the query data
Adds the extraction of sticky conditions to the useQueryFilters
composable. Because this change now outputs three variables, I've
changed the fuction return type to an object rather than an array for
ease of use. The newly added return value is an reactive array of
conditions that are sticky (i.e. can't easily be removed from the
overall query). See upstream django-filtering for complete feature
explanation.

This change is non-breaking, but not in use within the component
code that uses useQueryFilters. A later change provides logic to
utilize the full results of this change.