Enable users to edit sticky filters within the filtering app
Enables users to see and edit sticky filters with the filtering
app.

Note, this may not be the best solution to retrieving or displaying
the sticky filters. A better solution may be to embed behavior in the
Grouping and Condition classes, so that they become aware of there
capabilities and constraints. However, that solution is more of a
rewrite to the underlying classes. This solution works well enough,
though may not be as clear.