Fix the mock window.location for use with URL class
Fixes the mock location object to support _stringification_ via the
`toString` special method.

Fixes the mock location's href property to always have an
assignment. This fixes the case where `new URL(window.location)` needs
a base, not just the query data. There may be another way around this,
but it seems more straight forward to simply assign a default, because
it likely mimics the browser's behavior.