Change vue provide property name from index to listing
Changes the property name 'model-index-url' used in vue provide
definition to 'model-listing-url'. The change from 'index' to
'listing' is to try and unify around one term. I'm opting for the
django list (i.e. ListView) term over index.

The major project that uses this package uses the term 'index' for
listing pages, which is why the term crept in.