Use form submission on lozenge removal from listing entrypoint
The lozenges where previously being removed with a
window.location.push, which is limited in it's ability to reset the
state of the request. This for example becomes a problem with
pagination, which doesn't reset when one of the conditions is removed,
which should be the expected behavior.

The use of a form submission to the filtering form url ensures the
state resets and that the query is clean.