Fix submission when state change removes the to be submitted form
In the case where this is one lozenge that is removed, the form failed
to submit. This was because the prior fix prevented the submission,
which was put in place to ensure the form existed in order to submit.

The prior fix works when at least one lozenge or condition remains in
the query data. But once the next tick removes the last condition in
the query data, the element containing the form no longer rendered.

The form has been moved out of the visual element to ensure the
template ref will always be available.