Use the listing url provided by the backend
Use the listing url provided by the backend as is instead of trying to
build a url from what we know. Assume the backend knows best.

In shadowserver's particular usage we save filtering data in the
session to allow actions to happen between visits. The backend is able
to provide us with more information and continuity than the frontend
code for this particular case.