Allow grouped choices for choice lookup options
Django allows for grouped choices to be arrays of arrays which in
django forms translate to the use of the `optgroup` html tag. The
changes here enable that very same represenation.

Note, django uses a `[value, label]` structure for choices, until
grouping. Then it uses `[label, sub-values]`. As such, the code here
conforms to that higher design decision.