Allow for JSON data types in the condition value
Adjusts the condition value set to the the backend to be more than
just string type.

The initial usecase for this is a Yes/No to true/false values that
should maintain the Javascript/JSON boolean type rather than change it
to string. Othere cases would be for numbers and dates.